+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: fbrisson <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/22 10:57:41 by fbrisson          #+#    #+#             */
-/*   Updated: 2022/11/24 16:49:38 by fbrisson         ###   ########.fr       */
+/*   Updated: 2022/12/02 11:25:05 by fbrisson         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ int		ft_print_unsigned(unsigned int n);
 int		ft_printf(const char *s, ...);
 int		ft_printnbr(int n);
 int		ft_printpercent(void);
-int		ft_printptr(unsigned long long ptr);
+int		ft_printptr(unsigned long ptr);
 int		ft_printstr(char *s);
 int		ft_putchar(char c);
 int		ft_tri_selectif(va_list ptr_args, const char format);
